Removed .spec files.  Resolved most of the PR requests. Small improvements on code organisation.
