+++ src/app/models/projectTaskEntry.model.ts
@@ -1,0 +1,9 @@
+export interface ProjectTaskEntry {
+  id: number;
+  clientName: string;
+  projectName: string;
+  projectCode: string;
+  projectTask: string ;
+  durationInMin: number;
+  note: string;
+}
