+++ src/app/models/user.model.ts
@@ -1,0 +1,8 @@
+export interface User {
+  username: string;
+  firstName: string;
+  lastName: string;
+  role: string;
+  avatarURL: string;
+}
+
