--- src/app/pipes/format-minutes.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { FormatMinutesPipe } from './format-minutes.pipe';
-
-describe('FormatMinutesPipe', () => {
-  it('create an instance', () => {
-    const pipe = new FormatMinutesPipe();
-    expect(pipe).toBeTruthy();
-  });
-});
