--- src/app/role.directive.spec.ts
@@ -1,8 +1,0 @@
-import { RoleDirective } from './role.directive';
-
-describe('RoleDirective', () => {
-  it('should create an instance', () => {
-    const directive = new RoleDirective();
-    expect(directive).toBeTruthy();
-  });
-});
