+++ src/app/shared/material-design/material-design.module.ts
@@ -4,7 +4,7 @@ import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSidenavModule } from '@angular/material/sidenav';
-import {MatDialogModule, MatDialog, MatDialogRef} from '@angular/material/dialog';
+import { MatDialogModule } from '@angular/material';
 
 import { MatCardModule } from '@angular/material/card';
 import { MatListModule } from '@angular/material/list';
