+++ src/app/time-sheet/time-entry-dialogue/time-entry-dialogue.component.html
@@ -1,9 +1,6 @@
 <h1 mat-dialog-title>Hi </h1>
 <div mat-dialog-content>
   <p>{{data[0].projectName}}</p>
-  <!-- <mat-form-field>
-    <input matInput>
-  </mat-form-field> -->
 </div>
 <div mat-dialog-actions>
   <button mat-button (click)="onNoClick()">No Thanks</button>
