+++ src/app/time-sheet/time-entry-dialogue/time-entry-dialogue.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
-import { ProjectTaskEntry } from '../../services/timesheet.service';
+import { ProjectTaskEntry } from '../../models/projectTaskEntry.model';
 
 
 @Component({
