--- src/app/time-sheet/timesheets.json
@@ -1,24 +1,0 @@
-[
-    {
-        "client": "Tacta",
-        "projectName": "Taskter - Time tracking" ,
-        "projectCode": "TASKTER-TIME",
-        "tasks": ["Design","Programming","Analaysis"]
-    },
-
-    {
-        "client": "Tacta",
-        "projectName": "Taskter - Time tracking" ,
-        "projectCode": "TASKTER-TIME",
-        "tasks": ["Design","Programming","Analaysis"]
-    },
-
-    {
-        "client": "Tacta",
-        "projectName": "Taskter - Time tracking" ,
-        "projectCode": "TASKTER-TIME",
-        "tasks": ["Design","Programming","Analaysis"]
-    }
-   
-   
-]
